Update week-01-murray-poem.md
Hi Anandita! This looks good! Just some minor issues with the image syntax. The way you are linking to the images right now has the address to the web page where the image appears, but to have an image appear you need to have the full image URL with the image name and file extension, also for markdown to know you're about to display an image, you want to use an exclamation '!' before the brackets [ ] that contain the image title.

In the Image example on the markdown page: https://guides.github.com/features/mastering-markdown/#examples you'll see that their example has https://octodex.github.com/images/yaktocat.png with "yaktocat.png" being the actual image filename.

Normally you can find the filename of an image you find on the web by right clicking on it and clicking "open image in new tab" which will give you the exact address of that image file in your address bar - from there you can just copy it and paste it as shown in the markdown example. 

I've fixed your first image to show you what the syntax should look like. For pinterest you normally have to visit the pinned page to get to the menu where you can open the image in its own tab to get the address. 

Let me know if you have questions!