Update week-03-response
Comments: That is an accurate response about the hidden marketing that is often apparent in social media imagery. Popular people can sell things that they don't even use just by being in the near vicinity of the product. Do you think that this face can be abused or taken advantage of negatively? Or is it just a matter of fact situation?